Разбил код на модули, подготовил пустые модули под задачи
