+++ js/util.js
@@ -1,0 +1,11 @@
+// Случайное число в диапозоне
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.trunc(Math.random() * (upper - lower + 1) + lower);
+  return result;
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomArrayElement, getRandomInteger};
